requirements.txt
click==8.1.3
colorama==0.4.6
Flask==2.2.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.10.0
MarkupSafe==2.1.1
numpy==1.23.4
tensorflow-cpu==2.10.0
nltk==3.7
Werkzeug==2.2.2
